+++ index.html
@@ -17,8 +17,6 @@
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Comic+Neue:ital,wght@0,400;0,700;1,400;1,700&display=swap" rel="stylesheet">
 
-  <!-- Link to the JavaScript file -->
-  <script src="script.js" defer></script>
 </head>
 <body>
   <header>
@@ -26,9 +24,9 @@
       <section class="nav-bar-1">
         <h1>Awesome Books</h1>
         <ul class="nav-bar-ul">
-          <li id='nav-list' onclick="NewBook.addClassDisplay('list')" class="border-left-none">List</li>
-          <li id='nav-new' onclick="NewBook.addClassDisplay('addNew')" class="active" >Add New</li>
-          <li id='nav-contact' onclick="NewBook.addClassDisplay('contact')">Contact</li>
+          <li id='nav-list' class="border-left-none">List</li>
+          <li id='nav-new' class="active" >Add New</li>
+          <li id='nav-contact'>Contact</li>
         </ul>
       </section>
       <section class="nav-bar-2">
@@ -49,7 +47,7 @@
       <form id="bookForm" action="#">
         <input id="input-title"  type="text" placeholder="Title" required>
         <input id="input-author" type="text" placeholder="Author" required>
-        <button onclick="NewBook.add()" id="add-book" type="submit" class="addbtn">Add</button>
+        <button id="add-book" type="submit" class="addbtn">Add</button>
       </form>
     </section>
 
@@ -68,5 +66,7 @@
   <footer>
     <p>Copyright</p>
   </footer>
+  <!-- Link to the JavaScript file -->
+  <script type="module" src="script.js"></script>
 </body>
 </html>
