+++ script.js
@@ -1,148 +1,48 @@
-// Classe NewBook and the constructor
-let collectionBooks;
-class NewBook {
-  constructor(title, author) {
-    this.title = title;
-    this.author = author;
-    // this.index = index;
-  }
-
-  // ADD book
-  static add() {
-    const newtitle = document.querySelector('#input-title').value;
-    const newauthor = document.querySelector('#input-author').value;
-
-    // If the title or the author is empty, then don't add the book
-    if (newtitle === '' || newauthor === '') {
-      return null;
-    }
-
-    // Counter to define the index in the colections books
-    // position += 1;
-
-    // Create a new Book
-    const bookInfo = new NewBook(newtitle, newauthor);
-    collectionBooks.push(bookInfo);
-
-    // Print in the html file
-    NewBook.printf(collectionBooks);
-    return collectionBooks;
-  }
-
-  // DELET book
-  static delet(input) {
-    // Delet the number by index position
-    collectionBooks.splice(input, 1);
-
-    // Upgrade the DOM
-    NewBook.printf(collectionBooks);
-
-    // Upgrade the Local Storage
-    localStorage.setItem('data', JSON.stringify(collectionBooks));
-
-    return collectionBooks;
-  }
-
-  // Print the info in to the browser
-  static printf(input) {
-    document.getElementById('container-book').innerHTML = input.map((items, index) => `
-    <div id="cards">
-    <p>"${items.title}" by ${items.author} </p>
-    <button class="buttonRemove" onclick="NewBook.delet(${index})">Remove</button>
-    </div>
-    `).join('');
-  }
-
-  static getBooks() {
-    let books = [];
-    if (localStorage.getItem('data')) {
-      books = JSON.parse(localStorage.getItem('data'));
-    }
-    return books;
-  }
-
-  // Get date
-  static sendDate() {
-    const hora = Array.from(Date());
-    let ordinal;
-
-    const dayWeek = hora.splice(0, 3).join(''); // día de la semana
-    const monthAndDay = hora.splice(1, 6).join(''); // Mes y día
-    const year = hora.splice(2, 4).join(''); // Año
-    const hour = hora.splice(3, 8).join(''); // Hora, minutos y segundos
-
-    const monthAndDayStrg = monthAndDay.substring(4, 6);
-
-    if (monthAndDayStrg === 1 || monthAndDayStrg === 21) {
-      ordinal = 'st';
-    } if (monthAndDayStrg === 2 || monthAndDayStrg === 22) {
-      ordinal = 'nd';
-    } if (monthAndDayStrg === 3 || monthAndDayStrg === 23) {
-      ordinal = 'rd';
-    }
-
-    if (monthAndDayStrg <= 20 && monthAndDayStrg > 3) {
-      ordinal = 'th';
-    } if (monthAndDayStrg >= 24 && monthAndDayStrg <= 31) {
-      ordinal = 'th';
-    }
-    document.getElementById('date').innerHTML = `${dayWeek}, ${monthAndDay}${ordinal} ${year}, ${hour}`;
-  }
-
-  static addClassDisplay(input) {
-    switch (input) {
-      case 'list':
-        document.getElementById('listContainer').style.display = '';
-        document.getElementById('formContainer').style.display = 'none';
-        document.getElementById('contactContainer').style.display = 'none';
-
-        document.getElementById('nav-list').classList.add('active');
-        document.getElementById('nav-new').classList.remove('active');
-        document.getElementById('nav-contact').classList.remove('active');
-        break;
-
-      case 'addNew':
-        document.getElementById('listContainer').style.display = 'none';
-        document.getElementById('formContainer').style.display = '';
-        document.getElementById('contactContainer').style.display = 'none';
-
-        document.getElementById('nav-list').classList.remove('active');
-        document.getElementById('nav-new').classList.add('active');
-        document.getElementById('nav-contact').classList.remove('active');
-        break;
-
-      case 'contact':
-        document.getElementById('listContainer').style.display = 'none';
-        document.getElementById('formContainer').style.display = 'none';
-        document.getElementById('contactContainer').style.display = '';
-
-        document.getElementById('nav-list').classList.remove('active');
-        document.getElementById('nav-new').classList.remove('active');
-        document.getElementById('nav-contact').classList.add('active');
-        break;
-
-      default:
-        break;
-    }
-  }
-}
+import { addClassDisplay } from './modules/display-sections.js';
+import {
+  add, delet, printf,
+} from './modules/books.js';
+import { getBooks, cleanInputs } from './modules/storage.js';
+import { DateTime } from './modules/luxon.js';
 
 // Call the function to add classes to the main sections
-NewBook.addClassDisplay('addNew');
+document.getElementById('nav-list').addEventListener('click', () => {
+  addClassDisplay('list');
+});
+
+// Get date
+const sendDate = () => {
+  const myDate = DateTime.now();
+  document.getElementById('date').innerHTML = myDate.toLocaleString(DateTime.DATETIME_MED_WITH_SECONDS);
+};
 
 // Call get date
-window.addEventListener('DOMContenLoaded', NewBook.sendDate());
-setInterval(NewBook.sendDate, 1000);
+window.addEventListener('DOMContenLoaded', sendDate());
+setInterval(sendDate, 1000);
 
-// Get information from LocalStorage in to browser
-collectionBooks = NewBook.getBooks();
-NewBook.printf(collectionBooks);
+document.getElementById('nav-new').addEventListener('click', () => {
+  addClassDisplay('addNew');
+});
 
-// Save info in LocalStorage
+document.getElementById('nav-contact').addEventListener('click', () => {
+  addClassDisplay('contact');
+});
+// Set the Add Books section as the main one.
+addClassDisplay('addNew');
+
+const keepBooks = getBooks();
+printf(keepBooks);
+
+document.getElementById('add-book').addEventListener('click', () => {
+  add();
+});
+
 document.querySelector('#bookForm').addEventListener('submit', () => {
-  localStorage.setItem('data', JSON.stringify(collectionBooks));
+  cleanInputs();
+});
 
-  // Clean inputs info
-  document.querySelector('#input-title').value = '';
-  document.querySelector('#input-author').value = '';
+const booksContainer = document.getElementById('container-book');
+booksContainer.addEventListener('click', (e) => {
+  const { id } = e.target;
+  delet(id);
 });
